pyproject.toml
[tool.poetry]
name = "miniogre"
version = "0.1.0"
description = "Generate Dockerfile and requirements.txt from project directory"
authors = ["Your Name <you@example.com>"]
include = [".env"]

[tool.poetry.dependencies]
python = "^3.8"
typer = "0.9.0"
docker = "^5.0.0"
python-dotenv = "^0.19.0"
openai = "1.12.0"
emoji = "^2.10.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
miniogre = "miniogre.main:app"